Creation of elementary classes for the Peach exporter
A new abstract class exporterPlugin was made in the core.

A new package PeachExporter, following the MVC standard, was made in netzob-plugins.
So it contains a view, a controller, a model and a globalMenuEntry.
The controller and the globalMenuEntry are standard.
The view is simple and needs to be adjust.
The model is simple too and needs to be strengthen and sharpen.