GTK+3 port: Fix pack_start/pack_end calls
In PyGTK 2.24 the pack_start/pack_end method of GtkHBox
and GtkVBox takes 4 arguments : 1 mandatory, 3 with a
default value. In PyGObject/GTK+ 3, these methods
takes 4 *mandatory* arguments. The pygi-convert.sh
script tries to add the mandatory arguments but
sometimes fails when there are nested parentheses.
This commit include fixes for corner cases
that cannot be handled by the pygi-convert.sh script.

Moreover, GtkTreeViewColumn also have a pack_start
method. This method only takes 2 arguments in PyGTK 2.24
and in GTK+ 3. The pygi-convert.sh script tries to detect
if the object on which the pack_start method is called is
a GtkTreeViewColumn by checking if the variable name
is "column" ! This is a totally unreliable method
which produces a lot of false negatives. This commit
includes fixes for cases where GtkTreeViewColumn were not
detected by the pygi-convert.sh script.