Move the "save before quit" dialog from "destroy" event to "delete-event".
When closing the main window, the user is asked to save his work before quitting. With this commit, the dialog is opened when the "delete-event" signal is sent to the window, instead of the "destroy" signal. This allows to the dialog window to be opened on top of the main window.