Update resources management to add plugins support.
This commit provides to plugins the possibility to have their own static ressources
and to easily retrieve its location. Methods getStaticResourcesPath and getPluginStaticResourcesPath
are added to the NetzobPlugin class which allows to retrieve resources path.

In addition, the definition of the current plugin is provided to the newly created
AbstractPluginController and AbstractPluginView parents of any controllers and views of plugins.

Setup definition file has been upgraded in order to be more OS-Friendly (uses os.path.join instead of string concatenation
for paths).

Locales (i18n) are moved from netzob roots to netzob static resources path next to ui, logging and other classical
ressources.