Analyze and detect the use of a deprecated C extension (IssueId #206).
When installing Netzob, the python's build target has been edited
in order to generated a random Binary IDentifier (BID) which is
inserted in the source code of the libs (through the use of macros) and
in the built python sources (through an extended command for setuptools).

The BID is now checked using the DepCheck.py file which include
a static class which provides the required functions to search, detect and
validate the imported C extensions.