Add a message table widget to display messages
Added a messages table widget which can display the messages of
a given symbol (see MessageTableView and MessageTableController).
The vocabulary perspective can contain as much messages tables
as wanted to display the messages of several symbols at once.
A new message table can be open using the "new window" button
of the symbol list toolbar and can be closed using the
close button on its upper right corner.