On startup load the last loaded project from the current workspace + PEP8 fixes.
Loading the last project must consider the case were no last loaded project exists
which brings the possibility to show netzob main interface whithout any project loaded.
To do so, a small fix is introduced in the source code.

To resume this ccommit includes : that last loaded project is declared as the current one on startup,
a bug is fixed which prevented to start when no project was selected. Some minor PEP8 fixes are also
included.