Introduces the new variables in netzob.
An error in the referenced variable is resolved:
We used to need a vocabulary in order to create a referenced variable (to find the pointed variable).
We now save only the identifier.

An error in the repeat variable is resolved
The implementation of mutable and random flags was not finished.