Corrects three bugs in variables:
First, random and mutable were strings wich values were "True" or "False". This is because they were simply get from the xml save.
Now their initialization is : random = xmlRoot.get("random") == "True".
This returns a boolean which is true when the string value is "True" and false elsewhere.
The difference bewteen the two versions in a debugging message can not be seen with a classic print(random), we must use a print(random == True) which is a bit more tricky. Especially if we do not expect such a mistake.

Secondly, in the getChildren() function of class AbstractNodeVariable, if random was true (a string like "False" works too, obviously) we did:
self.children = random.shuffle(self.children)
Or random.shuffle() do not return a shuffled version from self.children but an empty list. Actually the self.children list is shuffled as-is. So we replace this line by:
random.shuffle(self.children)

Thirdly, in the loadFromXML() function of AlternateVariable, the constructor of the variable was:
AlternateVariable(xmlID, xmlMutable, xmlRandom, xmlName, children) instead of:
AlternateVariable(xmlID, xmlName, xmlMutable, xmlRandom, children)
So all values (all strings at the moment) had absurd values.