Implement the "loadProjectFromFile" which loads a project using any XML file.
A project is declared using an XML definition. In order to import a project, we need
to verify the given file is valid and will be sucessfully parsed afterwards. This
can be done before the "import" operation by allowing to load a file (load = validate)
given its absolute path (outside the current workspace).