Solve a bunch of bugs on outputMessages and inputMessages in the abstraction layer and adapt the function abstract() to the new variable model.
Symbols and data (formatted in bitarray) were given to the gui as is. Of course, the gui could not handle this objects.
This commit use the function str() to solve it.

The abstract function was not adapted to the new variable model, it kept calls to learn() and compare() functions whereas these functions must only called by the read() and write() functions of the variable classes.
This commit patches it, the code seems more trivial (only calls to write() and read() are performed) because the deep variable management is in the new variables classes.