Resolve some import problems in the symbols and messages tree view and add a "reset variable" options.
They were some unresolved imports in both classes TreeMessageController and TreeSymbolController. This commit solves them.
It also adds a "reset variable" option to the "configure variation of field" entry of the contextual menu of a message.
This option allows the user to reset the variable of a field to the default variable of this one.