Add the getProgeny() function, solve a bug in variable token management and rewrite some debugging information
The getProgeny() is a variable method, it returns the variable on which it is called and all its successor (progeny).
The list of variable contained by a symbol used to be the list of all root variable of its symbol's field. This function, called on these root variables, returns all variables of the project.

The bug in variable token management was, that a readingToken was never set to True but at initialization. So in an alternate variable if one variable failed to read, the global result was False.
Now when a child variable of an alternate variable succeeded in a read access, it sets the readingToken's ok flag to True.

Some debugging information have been reset and sharpened.
Many message printings were modified in order not to show a bitarray value anymore but an hexadecimal one instead, which is much more readable.