Change name of variable that are reserved by Python and add a VariableIDTreeController.
Variables like iter or id are reserved by Python, I change these functions to aniter anid or _id.

I also add a new class : VariableIDTreeController which helps the user to select a variable ID by selecting a variable in a global variable tree.
This class is represented by a tree view of all variables, it is called by a click on the pointed ID field of the variable creation view.
A variable ID is selected on this tree view by a right click on the proper variable.

This is not really convenient and needs a little improvement.