The ABR now verifies the validity of the remote server and of the API key.
When submiting a bug report, the ABR verifies the status of the remote server; If its
up, the SSL certificate is verified and afterward the provided API Key.

If the remote server doesn't have a validate certificate, a dedicated dialog
is displayed in order to request to the user the next operation. It can continue
or abort the connection.