Allow to move messages using Drag And Drop while considering regex inconsistency (IssueID #292).
The targeted symbol might not have a regex valid with the new message. To manage
this special issue, a dialog box is displayed to the user to ask what he wants to do.
Clean the targeted regex, re-compute it or move the message to the new Trash Symbol.

A message can be drag and dropped to a different symbol.
In this case, a dedicated dialog is displayed and ask
if the user wants to cancel the move or if he wants to adapt the
regex of the target symbol.

In also includes the definition of the TrashSymol.