Add a new way to define data block : delimiters.
Each DataVariable's type can now be sized or not. If it is sized, it is restricted by a minimum and maximum size as before and usual.
But if it is unsized, it is defined between 0 characters and the first delimiter found.
Delimiters are sets of bits used to separate unsized fields.

This commit implements the delimiters way in all type class.