Implementation of the new variables processing matrix.
This matrix changes the way the variable are read and write. Three flags/attributes interfere with the elementary access in order to make up to 16 different ways to process each variable.
Actually there are many similar case, see the matrix.

This commits implements the basic tricks, transformation, and naming but it does not close this implementation. In fact some major tree-managing problems are still present in the code. They will be resolved in the following commits.