Active Custom Mathematics Filters (IssueId #297).
The user can clicks on a symbol and ask to create a new custom filter.
It will allow him to specify the python code which will be executed as its filter source code.
Once created, the filter is available for messages and fields.
The filter can be activated or deactivated.

The saving process is not yet provided hence a created filter cannot be retrieved when
netzob is restarted.