Fix a bug in the C regex application.
The C module of regex doesn't consider groups of static fields. Hence
when two fields are static and one after the other, the two fields are
merged into one by the C module.

To fix this issue, static fields are parsed and their values are compared
with the values stored in fields definitions.