Modify the learn mechanism of all node variables.
The aim of the node learning is to insert a new child where the reading access fails. The value of the new child is computed on the unread value and stops where the remaining variables start.
Node variable learning (read accesss) is:
	- aggregate variable learns when one of its child could not read. It inserts a pad variable between the previoulsy treated child, and the child it should have read.
	- alternate variable learns when no child happens to read successfully. It gets though a new child. This child is determined by the value from the start of the remaining read variable and the start of the first right brother of the alternate.
	- repeat variable simply adapts their minimum and maximum iterations to the actual number of time they have read their value.

The pad variable's value is computed with a sequential loop.
This learning tries to fix a read of a node variable, it did not try to save the whole variable tree.