Move the "save before quit" dialog from "destroy" event to "delete-event".
When closing the main window, the user is asked to save his work before quitting. With this commit, the dialog is opened when the "delete-event" signal is sent to the window, instead of the "destroy" signal. This allows to the dialog window to be opened on top of the main window.

Remark: this commit is nearly the same as commit:874b79bb4bc2f3410e20fbe976bcd9c4949b73a3