Modify the method used to synchronize relation variable to their pointed variable.
There is no change for direct relation variables (variables that just have to take the value of the pointed variable and compute their own value on it. It is the cas of pointed variable that are on a different tree than their pointed variable or on the same tree but to the right.)

Undirect relation variables have been modified:
First, on reading they compare their format and reserved some bits of the read value, on writing, they write the value 'TEMP' in the written message. And they bind themselves to their pointed variable.
Secondly, when the pointed variable is treated, it notifies its pointing-to variable. This one retrieve the pointed variable's value, compares it to the value it has read in a reading access.
In a writing access, it overwrites the 'TEMP' value in the written message by the actualized value.

The choppedValue of the processing token is replaced by a linkedValue which is a list of couple (ID, value written/read by the variable that has this ID).

Moreover, this commit removes a lot of unused functions as getChoppedValue, trivialCompareFormat, getToken...
It also gives each variable a current value which is the value we can see from the given variable if the tree is flattened.