Introduce the Signals Manager.
In order to link a set of actions as the requirements of another set of actions we use
signals. Different type of signals exits like SIG_PROJECT_OPEN which when emitted means a project
has been opened. Another symbol for example is the SIG_SYMBOLS_MULTIPLE_SELECTION which means
that multiple symbols have been selected.

The signals manager brings together emitter of signals with listeners of signals. In order to listen to
the emission of symbols a method must be attached to a specific set of Signals using the 'attach' method.