Replace getValue with getTokenValue in ComputedRelationVariable.
ComputedRelationVariable used to find the base value on which they compute their with the getValue function.
But the getValue function is implemented only for variable that have values (DataVariable for instance).
So for node variable, it caused a crash.

getTokenValue returns the subvalue of the final value (in the VariableProcessingToken) a variable is responsible for.