Activate and enhance the tokenChoppedIndex list.
This list is used to follow which part of the final value is due to which variable. This is very useful in case of ComputedRelationVariable pointing on NodeVariable (that do not have an inherent value on which the computation can be based).

Moreover it ensures that only a written value is taking into account.