Implement factor, offset, WordSizeRelationType and BinaryRelationType into AbstractRelationType.
Factor and offset are added as two new attributes of AbstractRelationType. Getters and setters are added accordingly.
This was done because factor and offset may be needed by all other kinds of relations.

The makeType() static function was modified to implement this new attributes and to implement the difference between BinarySizeRelationType and WordSizeRelationType.