Implement factor and offset RelationType attributes into ComputedRelationVariable.py.
A ComputedRelationVariable owns at least an AbstractRelationType. AbstractRelationTypes have now two new attributes: factor and offset. As types are not saved directly, it is the owning ComputedRelationVariable that has to manage the reading from and writing to project XML file of this attribute.

This commit modifies toXML() and loadFromXML() method of ComputedRelationVariable in order to save/load the new attributes factor and offset.