Change the way RepeatVariables manage their children.
Before: RepeatVariables have a children list that ought to have only one element: the child of the RepeatVariable.
Now: RepeatVariable have a child that is their default child and serves as references for others. And they have children, children are organized in a list and are clones of child.

The following commits will explain this choice. Briefly we can say that now when a RepeatVariable have to execute a treatment (read or write a value), for each iteration (between minIterations and maxIterations) of this treatment it generates a new clone and executes the given treatment on this clone.