Modify the way RepeatVariables execute a read access.
Before: RepeatVariables execute this read access on their child as many times as they can between minIterations and maxIterations. They save little information before each iteration in order to be able to go back if an iteration failed. The problem was that each iteration modify the same variable. So if, under a repetition, a size variable pointed to another variable, the size variable would have, on each of its iteration the same value: the size of the last iteration of the pointed variable.

Now: RepeatVariables clone their child and then execute the read access on this clone for each iteration. So nothing has to be saved. RepeatVariable are now AggregateVariable which children are clones of the RepeatVariable 'child' (kind of alpha/elder/reference child). It solves the problems exposed previously.